Move the CRPS check entirely into serve.
