+++ serve/server/scripts/compare_crps.py
@@ -24,10 +24,10 @@ produce statistically equivalent ensemble forecasts.
 
 Usage
 -----
-    python compare.py --forecast-a /path/to/ferroflux/forecast.zarr \\
-                      --forecast-b /path/to/python_serve/forecast.zarr \\
-                      --variables t2m,z500,tcwv \\
-                      --threshold 0.01
+    python compare_crps.py --forecast-a /path/to/ferroflux/forecast.zarr \\
+                           --forecast-b /path/to/python_serve/forecast.zarr \\
+                           --variables t2m,z500,tcwv \\
+                           --threshold 0.01
 
 Expected zarr layout
 --------------------
@@ -39,11 +39,9 @@ is a separate data variable in the dataset.
 Testing
 -------
 Requires physicsnemo and earth2studio installed with the statistics extra.
-Run from the recipes/eval/ directory::
+Run from the repository root::
 
-    python compare.py --forecast-a tests/data/forecast_a.zarr \\
-                      --forecast-b tests/data/forecast_b.zarr \\
-                      --variables t2m --threshold 0.05
+    python -m pytest test/serve/server/test_compare_crps.py -v
 """
 
 from __future__ import annotations
@@ -52,16 +50,11 @@ import argparse
 import sys
 import warnings
 from collections import OrderedDict
+from typing import Any
 
 import numpy as np
 import torch
 import xarray as xr
-from src.data_loading import (
-    build_lead_time_chunks,
-    load_prediction_chunk,
-    load_verification_chunk,
-    spatial_coords_from_dataset,
-)
 
 from earth2studio.data import GFS
 from earth2studio.statistics import crps
@@ -69,6 +62,169 @@ from earth2studio.statistics.weights import lat_weight
 
 warnings.filterwarnings("ignore", message="Unclosed client session")
 warnings.filterwarnings("ignore", message="Unclosed connector")
+
+_NON_SPATIAL = frozenset({"ensemble", "time", "lead_time", "batch"})
+
+CoordSystem = OrderedDict
+
+
+def spatial_coords_from_dataset(ds: xr.Dataset) -> CoordSystem:
+    """Extract spatial coordinate arrays from a prediction Dataset.
+
+    Parameters
+    ----------
+    ds : xr.Dataset
+        Opened prediction zarr store.
+
+    Returns
+    -------
+    CoordSystem
+        Spatial coordinate arrays (e.g. ``{lat: [...], lon: [...]}``,
+        or ``{x: [...], y: [...]}`` for non-lat/lon grids).
+    """
+    coords: CoordSystem = OrderedDict()
+    for dim in ds.dims:
+        if dim not in _NON_SPATIAL:
+            coords[dim] = ds.coords[dim].values
+    return coords
+
+
+def build_lead_time_chunks(
+    lead_times: np.ndarray,
+    chunk_size: int | None,
+) -> list[np.ndarray]:
+    """Partition lead times into chunks for memory-bounded processing.
+
+    Parameters
+    ----------
+    lead_times : np.ndarray
+        Full array of lead-time values.
+    chunk_size : int | None
+        Maximum number of lead times per chunk.  ``None`` or ``<= 0``
+        means no chunking (all lead times in one chunk).
+
+    Returns
+    -------
+    list[np.ndarray]
+        List of lead-time sub-arrays.
+    """
+    if chunk_size is None or chunk_size <= 0 or chunk_size >= len(lead_times):
+        return [lead_times]
+    return [
+        lead_times[i : i + chunk_size] for i in range(0, len(lead_times), chunk_size)
+    ]
+
+
+def load_prediction_chunk(
+    prediction_ds: xr.Dataset,
+    time: np.datetime64,
+    lead_times: np.ndarray,
+    variables: list[str],
+    device: torch.device,
+) -> tuple[torch.Tensor, CoordSystem]:
+    """Load a chunk of prediction data for one IC time and lead-time range.
+
+    Parameters
+    ----------
+    prediction_ds : xr.Dataset
+        Opened prediction zarr store.
+    time : np.datetime64
+        Initial-condition time to select.
+    lead_times : np.ndarray
+        Lead-time values for this chunk.
+    variables : list[str]
+        Variable names to load.
+    device : torch.device
+        Target device for the returned tensor.
+
+    Returns
+    -------
+    tuple[torch.Tensor, CoordSystem]
+        Tensor and coords with dimensions
+        ``(ensemble?, lead_time, variable, <spatial...>)``.
+    """
+    subset = prediction_ds[variables].sel(time=time, lead_time=lead_times)
+    da = subset.to_array(dim="variable")
+
+    has_ensemble = "ensemble" in da.dims
+    spatial_dims = [
+        d for d in da.dims if d not in {"variable", "ensemble", "lead_time"}
+    ]
+
+    if has_ensemble:
+        dim_order = ["ensemble", "lead_time", "variable"] + spatial_dims
+    else:
+        dim_order = ["lead_time", "variable"] + spatial_dims
+
+    da = da.transpose(*dim_order)
+    tensor = torch.from_numpy(da.values.copy()).to(device=device, dtype=torch.float32)
+
+    coords: CoordSystem = OrderedDict()
+    for dim in dim_order:
+        coords[dim] = np.array(da.coords[dim].values)
+    return tensor, coords
+
+
+def load_verification_chunk(
+    source: Any,
+    time: np.datetime64,
+    lead_times: np.ndarray,
+    variables: list[str],
+    spatial_coords: CoordSystem,
+    device: torch.device,
+) -> tuple[torch.Tensor, CoordSystem]:
+    """Load verification data aligned to a prediction chunk.
+
+    For each lead time, the valid time ``time + lead_time`` is computed and
+    fetched from the verification source.  The returned tensor uses
+    ``lead_time`` (not valid time) as its first dimension so that it aligns
+    with the prediction chunk.
+
+    Parameters
+    ----------
+    source
+        Verification data source (callable accepting times and variables).
+    time : np.datetime64
+        Initial-condition time.
+    lead_times : np.ndarray
+        Lead-time values for this chunk.
+    variables : list[str]
+        Variable names to load.
+    spatial_coords : CoordSystem
+        Spatial coordinate arrays (for building the output CoordSystem).
+    device : torch.device
+        Target device for the returned tensor.
+
+    Returns
+    -------
+    tuple[torch.Tensor, CoordSystem]
+        Tensor and coords with dimensions
+        ``(lead_time, variable, <spatial...>)``.
+    """
+    valid_times = time + lead_times
+    da = source(list(valid_times), list(variables))
+
+    # Align verification grid to the prediction's spatial coordinates.
+    spatial_dims = [d for d in da.dims if d not in {"time", "variable"}]
+    sel_kwargs = {
+        dim: xr.DataArray(spatial_coords[dim], dims=[dim])
+        for dim in spatial_dims
+        if dim in spatial_coords
+    }
+    if sel_kwargs:
+        da = da.sel(**sel_kwargs, method="nearest")
+
+    dim_order = ["time", "variable"] + spatial_dims
+    da = da.transpose(*dim_order)
+
+    tensor = torch.from_numpy(da.values.copy()).to(device=device, dtype=torch.float32)
+
+    coords: CoordSystem = OrderedDict()
+    coords["lead_time"] = lead_times
+    coords["variable"] = np.array(variables)
+    for dim in spatial_dims:
+        coords[dim] = spatial_coords[dim]
+    return tensor, coords
 
 
 def parse_args() -> argparse.Namespace:
